Added a file descriptor-based get function.
So it turns out that when you open the file and close it for each entry search, it no longer leaves the fd in the spot before the record because, you know, it closes the file. So now there's a second function that restores the old behavior.

lol who needs testing